PATCH: Refactor exports to use `createCtx` module instead of `index` for improved structure
